+++ packages/kokoas-client/src/pages/projRegisterV2/api/convertProjToForm.ts
@@ -1,174 +1,18 @@
-import { BuildingType, ICustgroups, IEmployees, IProjects, IProjtypes, RecordCancelStatus, TAgents, Territory } from 'types';
+import { BuildingType, ICustgroups, IEmployees, IProjects, RecordCancelStatus, Territory } from 'types';
 import { TForm } from '../schema';
 import { formatDataId } from 'libs';
 import format from 'date-fns/format';
 import parseISO from 'date-fns/parseISO';
 import { groupAgentsByType } from 'api-kintone/src/projects/helpers/groupAgentsByType';
-import { getDefaultEmployee } from '../form';
+import { convertProjAgentsToForm } from './convertProjAgentsToForm';
 
-interface IGetPersistentFieldsParams {
-  projRec: IProjects,
-  projTypeRec: IProjtypes | undefined,
+interface IConvertProjToFormParams {
   hasContract: boolean,
-}
-interface IConvertProjToFormParams extends IGetPersistentFieldsParams {
   employeeRecs: IEmployees[],
-  custGroupRec: ICustgroups
+  custGroupRec: ICustgroups,
+  projRec: IProjects,
 }
 
-
-/**
- * 最新設定に依存しない情報。
- * 
- * * 紹介料率
- * * 利益率
- * * 役割別紹介料率
- */
-const getPersistentFields = ({
-  projTypeRec,
-  projRec,
-  hasContract,
-}: IGetPersistentFieldsParams): Pick<TForm, 'commissionRate' | 'commRateByRole' | 'profitRate' | 'commRateByEmployee'> => {
-
-  const {
-    yumeCommFeeRate,
-    profitRate,
-    commRateByRoleList,
-    commRateByEmpList,
-  } = projTypeRec || {};
-
-  const {
-    commissionRate: commissionRateFromProj,
-    profitRate: profitRateFromProj,
-    commRateByRoleList: commRateByRoleFromProj,
-    commRateByEmpList: commRateByEmpFromProj,
-  } = projRec;
-
-  // Default to project records values
-  let parsedCommRate: string = commissionRateFromProj.value;
-  let parsedProfitRate: string = profitRateFromProj.value;
-
-  // 役職による紹介料率
-  const filteredCommRateByRole = commRateByRoleFromProj?.value
-    .filter(({ value: { role } }) => !!role.value);
-
-  const projHasCommRateByRole = filteredCommRateByRole.length > 0;
-
-  let parsedCommRateByRoles: TForm['commRateByRole'] | null = projHasCommRateByRole
-    ? filteredCommRateByRole
-      .map(({
-        value: { role, commRateByRole },
-      }) => ({
-        role: role.value,
-        rate: Number(commRateByRole.value),
-      }))
-    : [];
-
-  // 個別紹介料率
-  const filteredCommRateByEmployee = commRateByEmpFromProj?.value
-    .filter(({ value: { commEmpId } }) => !!commEmpId.value);
-
-  const projHasCommRateByEmployee = filteredCommRateByEmployee.length > 0;
-
-  let parseCommRateByEmployee: TForm['commRateByEmployee'] = filteredCommRateByEmployee
-    .map(({ value: { commEmpId, commRateByEmp, commEmpName, commEmpRole } }) => ({
-      empId: commEmpId.value,
-      commEmpId: commEmpId.value,
-      commEmpName: commEmpName.value,
-      commEmpRate: Number(commRateByEmp.value),
-      commEmpRole: commEmpRole.value,
-      //commEmpRole:  ,
-      //rate: Number(commRateByEmp.value),
-    }));
-
-  if (!hasContract) {
-    // If there's no contract, use Project Type's values
-    parsedCommRate = parsedCommRate || yumeCommFeeRate?.value || '';
-    parsedProfitRate = parsedProfitRate || profitRate?.value || '';
-    parsedCommRateByRoles = projHasCommRateByRole
-      ? parsedCommRateByRoles
-      : commRateByRoleList?.value
-        .filter(({ value: { role } }) => !!role.value)
-        .map(({
-          value: { role, commRateByRole },
-        }) => ({
-          role: role.value,
-          rate: Number(commRateByRole.value),
-        }))
-      || [];
-
-    parseCommRateByEmployee = projHasCommRateByEmployee
-      ? parseCommRateByEmployee
-      : commRateByEmpList?.value
-        .filter(({ value: { empId } }) => !!empId.value)
-        .map(({ value: { empId, commRateByEmp, empName, empRole } }) => ({
-          empId: empId.value,
-          commEmpId: empId.value,
-          commEmpName: empName.value,
-          commEmpRole: empRole.value,
-          commEmpRate: Number(commRateByEmp.value),
-        })) || [];
-
-  }
-
-  // If there's no value, set to null to avoid validation error
-  return {
-    commissionRate: parsedCommRate === '' ? 0 : Number(parsedCommRate),
-    profitRate: parsedProfitRate === '' ? 0 : Number(parsedProfitRate),
-    commRateByRole: parsedCommRateByRoles,
-    commRateByEmployee: parseCommRateByEmployee,
-  };
-};
-
-/**
- * KintoneのAGデータをスキーマに沿った成形に変換する
- */
-const convertAgentsToForm = ({
-  agType,
-  agents,
-  employeeRecs,
-  cgAgents,
-}: {
-  agents: IProjects['agents']['value'] | undefined,
-  agType: TAgents
-  employeeRecs: IEmployees[],
-  cgAgents: ICustgroups['agents'],
-}) => {
-
-  const result = agents
-    ?.filter(({ value: { agentId } }) => !!agentId.value)
-    .map(({
-      value: {
-        agentId,
-        empRole,
-        agentName,
-        agentType,
-      },
-    }) => {
-
-      const agentRec = employeeRecs.find(({ uuid: _empId }) => _empId.value === agentId.value);
-      const cgAgent = cgAgents.value.find(({ value: { employeeId } }) => employeeId.value === agentId.value)?.value;
-
-      return ({
-        ...getDefaultEmployee((agentType.value || cgAgent?.agentType.value || agType) as TAgents),
-        empId: agentId.value,
-        empRole: empRole?.value || agentRec?.役職.value || '',
-        empName: agentName.value || cgAgent?.employeeName.value || agentRec?.文字列＿氏名.value || '',
-      });
-    }) || [];
-
-
-  const hasLastValue = !!result.at(-1)?.empId;
-
-  // add field if empty or last value is not empty, up to 2
-  if (!result.length || (hasLastValue && result.length < 2)) {
-    result.push(getDefaultEmployee(agType));
-  }
-
-
-  return result;
-
-};
 
 
 /***********************************
@@ -177,8 +21,6 @@ const convertAgentsToForm = ({
  **********************************/
 export const convertProjToForm = ({
   projRec,
-  projTypeRec,
-  hasContract,
   employeeRecs,
   custGroupRec,
 }: IConvertProjToFormParams): Partial<TForm> => {
@@ -263,19 +105,19 @@ export const convertProjToForm = ({
       .split(',')
       .filter(Boolean) as RecordCancelStatus[],
 
-    yumeAG: convertAgentsToForm({
+    yumeAG: convertProjAgentsToForm({
       agents: yumeAG,
       cgAgents: cgAgents,
       agType: 'yumeAG',
       employeeRecs,
     }),
-    cocoAG: convertAgentsToForm({
+    cocoAG: convertProjAgentsToForm({
       agents: cocoAG,
       cgAgents: cgAgents,
       agType: 'cocoAG',
       employeeRecs,
     }),
-    cocoConst: convertAgentsToForm({
+    cocoConst: convertProjAgentsToForm({
       agents: cocoConst,
       cgAgents: cgAgents,
       agType: 'cocoConst',
@@ -326,12 +168,6 @@ export const convertProjToForm = ({
     planApplicationDate: planApplicationDate.value ? parseISO(planApplicationDate.value) : null,
     paymentMethod: paymentMethod.value,
 
-    // Persistent fields
-    ...getPersistentFields({
-      projRec,
-      projTypeRec,
-      hasContract,
-    }),
 
     // 店舗情報
     storeId: storeId.value || cgStoreId.value,
